Move sync_l1 to common
