+++ crates/batch-prover/src/metrics.rs
@@ -11,6 +11,8 @@ pub struct BatchProverMetrics {
     pub current_l2_block: Gauge,
     #[metric(describe = "The duration of processing a single soft confirmation")]
     pub process_soft_confirmation: Histogram,
+    #[metric(describe = "The duration of scanning and processing a single L1 block")]
+    pub scan_l1_block: Histogram,
 }
 
 /// Batch prover metrics
